fix C++11 narrowing conversion error

+++ c_src/utpdrv.cc
@@ -116,7 +116,7 @@ static ErlDrvEntry drv_entry = {
     0,
     0,
     0,
-    ERL_DRV_EXTENDED_MARKER,
+    static_cast<int>(ERL_DRV_EXTENDED_MARKER),
     ERL_DRV_EXTENDED_MAJOR_VERSION,
     ERL_DRV_EXTENDED_MINOR_VERSION,
     ERL_DRV_FLAG_USE_PORT_LOCKING,
